requirements.txt
gevent >= 1.1rc2
requests >= 2.3.0
pyyaml
six
future
flask-restful
flask-basicauth
gevent-socketio
configure
yapsy